Use latest optiopay/kafka + update compatability
I noticed the health check not timing out under certain conditions and
traced it to the optiopay/kafka client library. That issue has since
been fixed in https://github.com/optiopay/kafka/pull/68. I've updated
the library to the latest version and also added all the new versions of
Kafka into the compatibility spec. I had to remove pre 0.10 Kafka
because it seems to not work with the new client library. I suspect this
is because it was in the 0.10 that an API protocol was added:
https://cwiki.apache.org/confluence/display/KAFKA/KIP-35+-+Retrieving+protocol+version